Fix simplifiing of filters like [field => '<>' => []].
For filters like [field => 'ANY OPERATOR' => []] the _field_to_sql
method returned WHERE ((NULL)) independently of OPERATOR.
It's correct behaviour if the operator is '=' or 'IN', but it's
incorrect and opposite to expected one if the operator is '<>' or
'NOT IN'. In this case WHERE ((1)) should be returned.
This commit fixes the problem.